+++ examples/dialog_box/test_1_base_box.py
@@ -1,6 +1,10 @@
 import pytest
 
-from hardpy import DialogBox, ImageComponent, run_dialog_box
+from hardpy import (
+    DialogBox,
+    ImageComponent,
+    run_dialog_box,
+)
 
 pytestmark = pytest.mark.module_name("Base dialog box")
 
@@ -16,8 +20,10 @@ def test_base_dialog_box():
         title_bar="Operator check",
         dialog_text="Press the Confirm button",
     )
-    response = run_dialog_box(dbx)
-    assert response
+    result = run_dialog_box(dbx)
+    assert result
+    assert result.pass_fail_result is None
+    assert result.widget_result is True
 
 
 @pytest.mark.case_name("Base dialog box with image")
@@ -27,8 +33,35 @@ def test_base_dialog_box_with_image():
         dialog_text="Press the Confirm button",
         image=ImageComponent(address="assets/test.png", width=100),
     )
-    response = run_dialog_box(dbx)
-    assert response
+    result = run_dialog_box(dbx)
+    assert result
+    assert result.pass_fail_result is None
+    assert result.widget_result is True
+
+
+@pytest.mark.case_name("Base dialog box with pass_fail")
+def test_base_dialog_box_with_pass_fail():
+    dbx = DialogBox(
+        title_bar="Operator check",
+        dialog_text="Press PASS or FAIL button",
+        pass_fail=True,
+    )
+    result = run_dialog_box(dbx)
+    assert result.is_pass
+    assert result.widget_result
+
+
+@pytest.mark.case_name("Base dialog box with image and pass_fail")
+def test_base_dialog_box_with_image_and_pass_fail():
+    dbx = DialogBox(
+        title_bar="Operator check",
+        dialog_text="Press PASS or FAIL button",
+        image=ImageComponent(address="assets/test.png", width=100),
+        pass_fail=True,
+    )
+    result = run_dialog_box(dbx)
+    assert result.is_pass
+    assert result.widget_result
 
 
 @pytest.mark.case_name("Empty test after")