+++ examples/dialog_box/test_2_input_field.py
@@ -26,6 +26,7 @@ test_html = """
 </html>
 """
 
+
 @pytest.mark.case_name("Text input")
 def test_text_input():
     dbx = DialogBox(
@@ -34,9 +35,10 @@ def test_text_input():
         widget=TextInputWidget(),
         font_size=18,
     )
-    response = run_dialog_box(dbx)
-    set_message(f"Entered text {response}")
-    assert response == "ok", "The entered text is not correct"
+    result = run_dialog_box(dbx)
+    assert result.pass_fail_result is None
+    assert result.widget_result == "ok"
+    set_message(f"Entered text {result.widget_result}")
 
 
 @pytest.mark.case_name("Text input with image")
@@ -48,9 +50,10 @@ def test_text_input_with_image():
         image=ImageComponent(address="assets/test.png", width=100),
         font_size=18,
     )
-    response = run_dialog_box(dbx)
-    set_message(f"Entered text {response}")
-    assert response == "ok", "The entered text is not correct"
+    result = run_dialog_box(dbx)
+    assert result.pass_fail_result is None
+    assert result.widget_result == "ok"
+    set_message(f"Entered text {result.widget_result}")
 
 
 @pytest.mark.case_name("Text input with html")
@@ -66,9 +69,25 @@ def test_text_input_with_html():
         ),
         font_size=18,
     )
-    response = run_dialog_box(dbx)
-    set_message(f"Entered text {response}")
-    assert response == "ok", "The entered text is not correct"
+    result = run_dialog_box(dbx)
+    assert result.pass_fail_result is None
+    assert result.widget_result == "ok"
+    set_message(f"Entered text {result.widget_result}")
+
+
+@pytest.mark.case_name("Text input with pass_fail")
+def test_text_input_with_pass_fail():
+    dbx = DialogBox(
+        dialog_text="Type 'ok' and press PASS or FAIL button",
+        title_bar="Example of text input with pass/fail",
+        widget=TextInputWidget(),
+        font_size=18,
+        pass_fail=True,
+    )
+    result = run_dialog_box(dbx)
+    assert result.is_pass
+    assert result.widget_result == "ok"
+    set_message(f"Pass/Fail: {result.is_pass}, Entered text: {result.widget_result}")
 
 
 @pytest.mark.case_name("Numeric input")
@@ -79,9 +98,10 @@ def test_num_input():
         title_bar="Example of entering a number",
         widget=NumericInputWidget(),
     )
-    response = int(run_dialog_box(dbx))
-    set_message(f"Entered number {response}")
-    assert response == test_num, "The entered number is not correct"
+    result = run_dialog_box(dbx)
+    assert result.pass_fail_result is None
+    assert int(result.widget_result) == test_num
+    set_message(f"Entered number {result.widget_result}")
 
 
 @pytest.mark.case_name("Numeric input with image")
@@ -93,9 +113,10 @@ def test_num_input_with_image():
         widget=NumericInputWidget(),
         image=ImageComponent(address="assets/test.png", width=100),
     )
-    response = int(run_dialog_box(dbx))
-    set_message(f"Entered number {response}")
-    assert response == test_num, "The entered number is not correct"
+    result = run_dialog_box(dbx)
+    assert result.pass_fail_result is None
+    assert int(result.widget_result) == test_num
+    set_message(f"Entered number {result.widget_result}")
 
 
 @pytest.mark.case_name("Numeric input with html")
@@ -107,6 +128,22 @@ def test_num_input_with_html():
         widget=NumericInputWidget(),
         html=HTMLComponent(html=test_html, is_raw_html=True, width=50),
     )
-    response = int(run_dialog_box(dbx))
-    set_message(f"Entered number {response}")
-    assert response == test_num, "The entered number is not correct"
+    result = run_dialog_box(dbx)
+    assert result.pass_fail_result is None
+    assert int(result.widget_result) == test_num
+    set_message(f"Entered number {result.widget_result}")
+
+
+@pytest.mark.case_name("Numeric input with pass_fail")
+def test_num_input_with_pass_fail():
+    test_num = 123
+    dbx = DialogBox(
+        dialog_text=f"Enter the number {test_num} and press PASS or FAIL button",
+        title_bar="Example of entering a number with pass/fail",
+        widget=NumericInputWidget(),
+        pass_fail=True,
+    )
+    result = run_dialog_box(dbx)
+    assert result.is_pass
+    assert int(result.widget_result) == test_num
+    set_message(f"Pass/Fail: {result.is_pass}, Entered number: {result.widget_result}")